worker/requirements.txt
asyncio
grpcio
grpcio-tools
rawpy
PyExifTool
requests
Pillow